add EditServlet realization

+++ src/main/java/ua/nure/kn/dobriak/usermanagement/User.java
@@ -48,7 +48,7 @@ public class User {
 	public Integer getId() {
 		return id;
 	}
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
